fixed whitelist_check
in edit distance check, self.sender is modified to become the sender domain instead, which messes with whitelist check (as whitelist check requires the full sender address).
So i shifted whitelist check to be done first before distance check (in final risk score) and it works properly now. do take note that the output on /emails also shows the sender being the domain instead of the full sender address.
In whitelist check, the whitelist file has been updated to the new whitelist file (as of 28 sep 25 1330H) as formatting of the csv file has been fixed